// frontend/src/app/generated/model/modelsResourcePermissions.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModelsPolicyRule } from './modelsPolicyRule';

export interface ModelsResourcePermissions {
  allowedAll?: boolean;
  allowedInstances?: Array<string>;
  /**
   * Records which rule allowed the access
   */
  matchedRule?: ModelsPolicyRule;
}